+++ laboratory-problems/serial_dilution_factor_diluent_numeric.py
@@ -17,9 +17,12 @@ df_ratios = [
 #==================================================
 #==================================================
 def question_text(volume, df1, df2):
-	question = "<p>Using a previous diluted sample at DF={0}, ".format(df1)
-	question += "create a new dilution with a final dilution of DF={0} and a total volume of {1}&nbsp;&mu;L.</p>".format(df2, volume)
-	question += "<p>What volume of diluent in microliters (&mu;L) do you add to the aliquoted sample?</p>"
+	volume_text = f"<span style='font-family: monospace;'>{volume} &mu;L</span>"
+	question = (
+		f"<p>Using a previous diluted sample at DF={df1}, "
+		f"create a new dilution with a final dilution of DF={df2} and a total volume of {volume_text}.</p>"
+		"<p>What volume of diluent in microliters (&mu;L) do you add to the aliquoted sample?</p>"
+	)
 	return question
 
 #==================================================
@@ -35,29 +38,22 @@ def df_ratio_to_values(df_ratio):
 
 #==================================================
 #==================================================
-def write_question_batch(start_num, args):
-	question_list = []
-	question_num = start_num
-	for df_ratio in df_ratios:
-		if df_ratio[1] < 3:
-			continue
-		volume, df1, df2 = df_ratio_to_values(df_ratio)
-		q = question_text(volume, df1, df2)
-		aliquot = volume * df_ratio[1] / df_ratio[0]
-		diluent = volume - aliquot
-		answer = diluent
-		tolerance = 0.9
-		bbf = bptools.formatBB_NUM_Question(question_num, q, answer, tolerance)
-		question_list.append(bbf)
-		question_num += 1
-	return question_list
+def write_question(N: int, args) -> str:
+	valid_ratios = [ratio for ratio in df_ratios if ratio[1] >= 3]
+	df_ratio = random.choice(valid_ratios)
+	volume, df1, df2 = df_ratio_to_values(df_ratio)
+	q = question_text(volume, df1, df2)
+	aliquot = volume * df_ratio[1] / df_ratio[0]
+	diluent = volume - aliquot
+	tolerance = 0.9
+	bbf = bptools.formatBB_NUM_Question(N, q, diluent, tolerance)
+	return bbf
 
 #==================================================
 #==================================================
 def parse_arguments():
 	parser = bptools.make_arg_parser(
-		description='Generate serial dilution factor diluent numeric questions.',
-		batch=True
+		description='Generate serial dilution factor diluent numeric questions.'
 	)
 	args = parser.parse_args()
 	return args
@@ -67,8 +63,7 @@ def parse_arguments():
 def main():
 	args = parse_arguments()
 	outfile = bptools.make_outfile()
-	questions = bptools.collect_question_batches(write_question_batch, args)
-	bptools.write_questions_to_file(questions, outfile)
+	bptools.collect_and_write_questions(write_question, args, outfile)
 
 #==================================================
 #==================================================